googleyolo/googleyolo/requirements.txt
Django==2.0.4
google-auth==1.4.1
